requirements.txt
# ============================================================================
# Medical VQA - Knowledge-Guided Explainable Transformer
# Requirements File
# ============================================================================

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers & PEFT
transformers>=4.37.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99

# Qwen-VL specific
qwen-vl-utils>=0.0.4
tiktoken>=0.5.0

# DeepSpeed for efficient training
deepspeed>=0.12.0

# Knowledge Encoders
# Note: Install scispacy models separately:
# pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_lg-0.5.1.tar.gz
scispacy>=0.5.1
spacy>=3.7.0

# Explainability
grad-cam>=1.4.8
captum>=0.7.0
shap>=0.44.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1
pydicom>=2.4.0
SimpleITK>=2.3.0

# Metrics & Evaluation
nltk>=3.8.1
rouge-score>=0.1.2
sacrebleu>=2.3.0
scikit-learn>=1.3.0

# Web Application
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.26.0

# API Schemas
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
omegaconf>=2.3.0

# Logging & Monitoring
tensorboard>=2.15.0
wandb>=0.16.0
loguru>=0.7.2
tqdm>=4.66.0

# Data Handling
pandas>=2.1.0
numpy>=1.24.0
datasets>=2.16.0
pyarrow>=14.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utilities
einops>=0.7.0
safetensors>=0.4.0
huggingface-hub>=0.20.0
filelock>=3.13.0
requests>=2.31.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0
